python/django/pyproject.toml
[project]
name = "django-app"
version = "1.0.0"
description = "Django Application"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-environ>=0.11.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
    "django-redis>=5.4.0",
    "celery>=5.3.0",
    "django-celery-beat>=2.5.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "django-extensions>=3.2.0",
    "psutil>=5.9.0",
]
requires-python = ">=3.11"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "django-stubs>=4.2.0",
    "factory-boy>=3.3.0",
    "django-debug-toolbar>=4.2.0",
    "ipython>=8.16.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
addopts = "--tb=short --strict-markers --strict-config"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
] 